Add files via upload
The .ipynb file shows the 5 algorithms and the chain matrix algorithm as well. The 5 algorithms are compared to each others' runtimes for different scenarios. The chain matrix multiplication algorithm is compared with previous algorithms as well to check efficiency. All comparisons show algorithms against recurrence relations as well. Graphs shown.